+++ lib/zen/spec/helper.rb
@@ -1,10 +1,10 @@
 require 'ramaze'
 require 'stringio'
 
-Ramaze.setup(:verobse => false) do
-  gem 'capybara', ['= 1.0.1']
-  gem 'bacon'   , ['= 1.1.0']
-  gem 'webmock' , ['= 1.6.4']
+Ramaze.setup(:verbose => false) do
+  gem 'capybara', ['>= 1.1.1']
+  gem 'bacon'   , ['>= 1.1.0']
+  gem 'webmock' , ['>= 1.6.4']
 end
 
 require 'capybara/dsl'